feat(price): aligned UI for price on Product Page
