+++ packages/falcon-ecommerce-uikit/src/Product/ProductQuery.tsx
@@ -1,6 +1,35 @@
 import gql from 'graphql-tag';
 import { Query } from '../Query';
 
+export type ProductPrice = {
+  regular: string;
+  special: string;
+  minTier: string;
+  min: string;
+  max: string;
+};
+
+export type Product = {
+  id: string;
+  sku: string;
+  name: string;
+  image?: string;
+  urlPath: string;
+  thumbnail?: string;
+  price: ProductPrice;
+  priceType: string;
+  minPrice: string;
+  maxPrice: string;
+  currency: string;
+  description: string;
+  stock: any; // Stock
+  type: string;
+  configurableOptions: []; // [ConfigurableProductOption]
+  bundleOptions: []; // [BundleProductOption]
+  gallery: []; // [GalleryEntry]
+  seo: any; // ProductSeo
+};
+
 export const GET_PRODUCT = gql`
   query Product($id: String!, $path: String!) {
     product(id: $id) {
@@ -8,7 +37,11 @@ export const GET_PRODUCT = gql`
       sku
       name
       description
-      price
+      price {
+        regular
+        special
+        minTier
+      }
       currency
       gallery {
         full
@@ -38,7 +71,7 @@ export const GET_PRODUCT = gql`
   }
 `;
 
-export class ProductQuery extends Query<any> {
+export class ProductQuery extends Query<{ product: Product }> {
   static defaultProps = {
     query: GET_PRODUCT
   };
