+++ packages/falcon-ecommerce-uikit/src/ProductsList/ProductsListQuery.tsx
@@ -1,34 +1,6 @@
 import gql from 'graphql-tag';
 import { Query } from '../Query';
-
-export type ProductPrice = {
-  regular: string;
-  special: string;
-  minTier: string;
-  min: string;
-  max: string;
-};
-
-export type Product = {
-  id: string;
-  sku: string;
-  name: string;
-  image?: string;
-  urlPath: string;
-  thumbnail?: string;
-  price: ProductPrice;
-  priceType: string;
-  minPrice: string;
-  maxPrice: string;
-  currency: string;
-  description: string;
-  stock: any; // Stock
-  type: string;
-  configurableOptions: []; // [ConfigurableProductOption]
-  bundleOptions: []; // [BundleProductOption]
-  gallery: []; // [GalleryEntry]
-  seo: any; // ProductSeo
-};
+import { Product } from '../Product/ProductQuery';
 
 export const GET_PRODUCTS_LIST = gql`
   query Products {
